both mocke_environment together
