+++ services/dynamic-sidecar/tests/unit/conftest.py
@@ -4,7 +4,6 @@
 
 import asyncio
 import logging
-from pathlib import Path
 from typing import AsyncIterable, AsyncIterator
 from unittest.mock import AsyncMock
 
@@ -12,9 +11,7 @@ import pytest
 from aiodocker.volumes import DockerVolume
 from async_asgi_testclient import TestClient
 from fastapi import FastAPI
-from pytest import MonkeyPatch
 from pytest_mock import MockerFixture
-from pytest_simcore.helpers.utils_envs import EnvVarsDict, setenvs_as_envfile
 from simcore_service_dynamic_sidecar.core.application import AppState, create_app
 from simcore_service_dynamic_sidecar.core.docker_utils import docker_client
 from simcore_service_dynamic_sidecar.core.shared_handlers import (
@@ -26,15 +23,6 @@ from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
 
 logger = logging.getLogger(__name__)
-
-
-@pytest.fixture
-def mock_environment_with_envdevel(
-    monkeypatch: MonkeyPatch, project_slug_dir: Path
-) -> EnvVarsDict:
-    env_file = project_slug_dir / ".env-devel"
-    envs = setenvs_as_envfile(monkeypatch, env_file.read_text())
-    return envs
 
 
 #
